requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
h5py>=3.8.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-image>=0.20.0
einops>=0.6.0
timm>=0.9.0
matplotlib>=3.7.0
pandas>=2.0.0
